--- a/lib/wallet/src/Cardano/Wallet/Shelley/Transaction.hs
+++ b/lib/wallet/src/Cardano/Wallet/Shelley/Transaction.hs
@@ -277,6 +277,7 @@ import Ouroboros.Network.Block
 
 import qualified Cardano.Api as Cardano
 import qualified Cardano.Api.Byron as Byron
+import qualified Cardano.Api.Extra as Cardano
 import qualified Cardano.Api.Shelley as Cardano
 import qualified Cardano.Chain.Common as Byron
 import qualified Cardano.Crypto as CC
@@ -1256,6 +1257,10 @@ assignScriptRedeemers pparams ti utxo redeemers tx =
 
     systemStart = getSystemStart ti
 
+    pparams' = Cardano.unbundleLedgerShelleyBasedProtocolParams
+        (shelleyBasedEra @era)
+        pparams
+
     -- | Assign redeemers with null execution units to the input transaction.
     --
     -- Redeemers are determined from the context given to the caller via the
@@ -1336,9 +1341,6 @@ assignScriptRedeemers pparams ti utxo redeemers tx =
         -> Either ErrAssignRedeemers
             (Map Alonzo.RdmrPtr (Either ErrAssignRedeemers Alonzo.ExUnits))
     evaluateExecutionUnitsAlonzo indexedRedeemers alonzoTx = do
-        let pparams' = Cardano.toLedgerPParams
-                Cardano.ShelleyBasedEraAlonzo
-                    (Cardano.unbundleProtocolParams pparams)
         let costs = toCostModelsAsArray
                 (Alonzo.unCostModels $ Alonzo._costmdls pparams')
         let res = evaluateTransactionExecutionUnits
@@ -1361,9 +1363,6 @@ assignScriptRedeemers pparams ti utxo redeemers tx =
         -> Either ErrAssignRedeemers
             (Map Alonzo.RdmrPtr (Either ErrAssignRedeemers Alonzo.ExUnits))
     evaluateExecutionUnitsBabbage indexedRedeemers babbageTx = do
-        let pparams' = Cardano.toLedgerPParams
-                Cardano.ShelleyBasedEraBabbage
-                    (Cardano.unbundleProtocolParams pparams)
         let costs = toCostModelsAsArray
                 (Alonzo.unCostModels $ Babbage._costmdls pparams')
 
@@ -1387,9 +1386,6 @@ assignScriptRedeemers pparams ti utxo redeemers tx =
         -> Either ErrAssignRedeemers
             (Map Alonzo.RdmrPtr (Either ErrAssignRedeemers Alonzo.ExUnits))
     evaluateExecutionUnitsConway indexedRedeemers conwayTx = do
-        let pparams' = Cardano.toLedgerPParams
-                Cardano.ShelleyBasedEraConway
-                    (Cardano.unbundleProtocolParams pparams)
         let costs = toCostModelsAsArray
                 (Alonzo.unCostModels $ Conway._costmdls pparams')
 
@@ -1472,7 +1468,9 @@ assignScriptRedeemers pparams ti utxo redeemers tx =
 
     -- | Finally, calculate and add the script integrity hash with the new
     -- final redeemers, if any.
-    addScriptIntegrityHashAlonzo :: AlonzoTx -> AlonzoTx
+    addScriptIntegrityHashAlonzo
+        :: era ~ Cardano.AlonzoEra
+        => AlonzoTx -> AlonzoTx
     addScriptIntegrityHashAlonzo alonzoTx =
         let wits  = Alonzo.wits alonzoTx
             langs =
@@ -1484,17 +1482,15 @@ assignScriptRedeemers pparams ti utxo redeemers tx =
         in alonzoTx
             { Alonzo.body = (Alonzo.body alonzoTx)
                 { Alonzo.scriptIntegrityHash = Alonzo.hashScriptIntegrity
-                    (Set.fromList $ Alonzo.getLanguageView
-                        (Cardano.toLedgerPParams
-                            Cardano.ShelleyBasedEraAlonzo
-                                (Cardano.unbundleProtocolParams pparams))
-                        <$> langs)
+                    (Set.fromList $ Alonzo.getLanguageView pparams' <$> langs)
                     (Alonzo.txrdmrs wits)
                     (Alonzo.txdats wits)
                 }
             }
 
-    addScriptIntegrityHashBabbage :: BabbageTx -> BabbageTx
+    addScriptIntegrityHashBabbage
+        :: era ~ Cardano.BabbageEra
+        => BabbageTx -> BabbageTx
     addScriptIntegrityHashBabbage babbageTx =
         let wits  = Alonzo.wits babbageTx
             langs =
@@ -1506,17 +1502,16 @@ assignScriptRedeemers pparams ti utxo redeemers tx =
         in babbageTx
             { Babbage.body = (Babbage.body babbageTx)
                 { Babbage.scriptIntegrityHash = Alonzo.hashScriptIntegrity
-                    (Set.fromList $ Alonzo.getLanguageView
-                        (Cardano.toLedgerPParams
-                            Cardano.ShelleyBasedEraBabbage
-                                (Cardano.unbundleProtocolParams pparams))
+                    (Set.fromList $ Alonzo.getLanguageView pparams'
                         <$> langs)
                     (Alonzo.txrdmrs wits)
                     (Alonzo.txdats wits)
                 }
             }
 
-    addScriptIntegrityHashConway :: ConwayTx -> ConwayTx
+    addScriptIntegrityHashConway
+        :: era ~ Cardano.ConwayEra
+        => ConwayTx -> ConwayTx
     addScriptIntegrityHashConway conwayTx =
         let wits  = Alonzo.wits conwayTx
             langs =
@@ -1528,10 +1523,7 @@ assignScriptRedeemers pparams ti utxo redeemers tx =
         in conwayTx
             { Conway.body = (Conway.body conwayTx)
                 { Conway.scriptIntegrityHash = Alonzo.hashScriptIntegrity
-                    (Set.fromList $ Alonzo.getLanguageView
-                        (Cardano.toLedgerPParams
-                            Cardano.ShelleyBasedEraConway
-                                (Cardano.unbundleProtocolParams pparams))
+                    (Set.fromList $ Alonzo.getLanguageView pparams'
                         <$> langs)
                     (Alonzo.txrdmrs wits)
                     (Alonzo.txdats wits)