- [x] Replace all non-unit test calls to `Cardano.toLedgerPParams` with the much cheaper `Cardano.unbundleLedgerShelleyBasedProtocolParams`.

### Comments

Suggestion to #3812 

<!-- Additional comments, links, or screenshots to attach, if any. -->

### Issue Number

<!-- Reference the Jira/GitHub issue that this PR relates to, and which requirements it tackles.
  Note: Jira issues of the form ADP- will be auto-linked. -->