@Anviking 2023-03-28
Running `cabal test cardano-wallet:unit --test-options='-j 8 --match "balanceTransaction"'`

This branch:

```
Cardano.Wallet.Shelley.Transaction
  balanceTransaction
    doesn't balance transactions with existing 'totalCollateral' [✔] (661ms)
      +++ OK, passed 10 tests; 42 discarded.
    doesn't balance transactions with existing 'returnCollateral' [✔] (402ms)
      +++ OK, passed 10 tests; 25 discarded.
    produces valid transactions or fails [✔] (72290ms)
      +++ OK, passed 1000 tests:
       0.3% partial tx had zero ada outputs

      41.1% input resolution conflicts
      19.0% output below minCoinValue
      16.8% conflicting networks
       9.2% missing tokens
       8.6% success
       1.8% missing coin
       1.4% missing coin and tokens
       1.2% existing collateral
       0.9% maxTxSize limit exceeded
    bootstrap witnesses
      coin-selection's size estimation == balanceTx's size estimation [✔] (904ms)
        +++ OK, passed 100 tests.
      balanceTx's size estimation >= measured serialized size [✔] (985ms)
        +++ OK, passed 100 tests.

        overestimation/wit (100 in total):
        41% 9
        37% 2
         6% 4
         3% 0 (but with no wits)
         3% 11
         3% 5
         3% 6
         2% 7
         1% 10
         1% 8
    balance goldens
      testPParams [✔] (17ms)
        Golden and Actual output didn't change
      balanced binaries
        pingPong_2 [✔] (12ms)
          Golden and Actual output didn't change
      results when varying wallet balance (1 UTxO)
        pingPong_1 [✔] (29ms)
          Golden and Actual output didn't change
        pingPong_2 [✔] (83ms)
          Golden and Actual output didn't change
        delegate [✔] (27ms)
          Golden and Actual output didn't change
        1ada-payment [✔] (54ms)
          Golden and Actual output didn't change
    posAndNegFromCardanoValue
      roundtrips with toCardanoValue [✔]
        +++ OK, passed 100 tests.
    change address generation
      assigns change addresses as expected [✔] (9ms)
      returns s' corresponding to which addresses were used [✔]
    increases zero-ada outputs to minimum [‐]
      # PENDING: Needs correct mock PParams for Babbage
    effect of txMaxSize on coin selection
      tries to select 2x the payment amount [✔] (4ms)
      falls back to 1x if out of space [✔] (6ms)
      otherwise fails with ErrBalanceTxMaxSizeLimitExceeded [✔] (12ms)
    when passed unresolved inputs
      fails with ErrBalanceTxUnresolvedTxIn [✔]
      with redeemers
        fails with ErrBalanceTxUnresolvedInputs [✔]
    when validity interval is too far in the future
      with some Plutus redeemers
        fails with TimeTranslationPastHorizon [✔] (1ms)
      with no redeemers
        succeeds at balancing [✔]
    when a redeemer is missing
      balancing succeeds (currently) [✔]
    when a redeemer points to an input that doesn't exist
      fails with ErrAssignRedeemersTargetNotFound [✔]

Finished in 75.5092 seconds, used 82.4280 seconds of CPU time
```

#3812 48c86f2c3fb1aea2f50b04abc5cada9734f7e395

```
Cardano.Wallet.Shelley.Transaction
  balanceTransaction
    doesn't balance transactions with existing 'totalCollateral' [✔] (899ms)
      +++ OK, passed 10 tests; 49 discarded.
    doesn't balance transactions with existing 'returnCollateral' [✔] (398ms)
      +++ OK, passed 10 tests; 29 discarded.
    produces valid transactions or fails [✔] (86118ms)
      +++ OK, passed 1000 tests:
       0.4% partial tx had zero ada outputs

      40.8% input resolution conflicts
      19.5% conflicting networks
      17.9% output below minCoinValue
       8.4% success
       7.5% missing tokens
       2.0% missing coin
       1.7% missing coin and tokens
       1.4% existing collateral
       0.8% maxTxSize limit exceeded
    bootstrap witnesses
      coin-selection's size estimation == balanceTx's size estimation [✔] (876ms)
        +++ OK, passed 100 tests.
      balanceTx's size estimation >= measured serialized size [✔] (949ms)
        +++ OK, passed 100 tests.

        overestimation/wit (100 in total):
        40% 9
        39% 2
        10% 4
         3% 5
         3% 8
         2% 0 (but with no wits)
         2% 7
         1% 10
    balance goldens
      testPParams [✔] (32ms)
        Golden and Actual output didn't change
      balanced binaries
        pingPong_2 [✔] (59ms)
          Golden and Actual output didn't change
      results when varying wallet balance (1 UTxO)
        pingPong_1 [✔] (1321ms)
          Golden and Actual output didn't change
        pingPong_2 [✔] (2847ms)
          Golden and Actual output didn't change
        delegate [✔] (1316ms)
          Golden and Actual output didn't change
        1ada-payment [✔] (1346ms)
          Golden and Actual output didn't change
    posAndNegFromCardanoValue
      roundtrips with toCardanoValue [✔]
        +++ OK, passed 100 tests.
    change address generation
      assigns change addresses as expected [✔] (24ms)
      returns s' corresponding to which addresses were used [✔]
    increases zero-ada outputs to minimum [‐]
      # PENDING: Needs correct mock PParams for Babbage
    effect of txMaxSize on coin selection
      tries to select 2x the payment amount [✔] (20ms)
      falls back to 1x if out of space [✔] (38ms)
      otherwise fails with ErrBalanceTxMaxSizeLimitExceeded [✔] (43ms)
    when passed unresolved inputs
      fails with ErrBalanceTxUnresolvedTxIn [✔]
      with redeemers
        fails with ErrBalanceTxUnresolvedInputs [✔]
    when validity interval is too far in the future
      with some Plutus redeemers
        fails with TimeTranslationPastHorizon [✔] (17ms)
      with no redeemers
        succeeds at balancing [✔] (15ms)
    when a redeemer is missing
      balancing succeeds (currently) [✔] (46ms)
    when a redeemer points to an input that doesn't exist
      fails with ErrAssignRedeemersTargetNotFound [✔] (16ms)

Finished in 96.3977 seconds, used 105.0684 seconds of CPU time
```

Looking particularly good is the speedup of the golden tests:

Golden | time before / ms | time after / ms | speedup
--                 | --     | --   | --   |
pingPong_1 | 1321 | 29 | 46x
pingPong_2 | 2847 | 83 | 34x
delegate | 1316 | 27 | 48x |
1-ada-payment | 1346 | 54 | 25x