@jonathanknowles 2023-03-29
@Anviking I did some performance testing here: https://github.com/input-output-hk/cardano-wallet/pull/3812#issuecomment-1488156169